backend/requirements.txt
Flask==2.2.3
Flask-SQLAlchemy==3.0.3
Flask-WTF==1.1.1
Flask-CORS==3.0.10
Werkzeug==2.2.3
selenium>=4.16.0
webdriver-manager>=4.0.0
requests>=2.28.2,<2.32.0
beautifulsoup4>=4.11.2
python-dotenv>=1.0.0
line-bot-sdk==2.4.2
schedule>=1.1.0
SQLAlchemy>=2.0.38
typing_extensions>=4.12.2
lxml>=4.9.3 